Split test_rb_utils into itself and test_clifford_utils
